Refactor API service layer and update imports
Removed lib/api-service.ts and split its functionality into new service modules: chat-service, dashboard-service, log-service, menu-service, permission-service, role-service, and user-service. Updated all relevant imports throughout the codebase to use the new service modules, improving code organization and maintainability.
