+++ components/features/users/add-user-button.tsx
@@ -3,7 +3,7 @@
 import { Dialog, Button, Flex, Text } from "@radix-ui/themes";
 import { useState, useEffect } from "react";
 import { createUser } from "@/services/user-service";
-import { getMenuBranches, getMenuRoles } from "@/lib/api-service";
+import { getMenuBranches, getMenuRoles } from "@/services/menu-service";
 import type { Branch, Role, CreateUserPayload } from "@/types";
 import {
   showCreateSuccess,
