+++ components/features/users/create-user-dialog.tsx
@@ -21,7 +21,7 @@ import {
 } from "@/components/ui/select";
 import { Users, FileText, Mail, Lock, Briefcase } from "lucide-react";
 import { createUser } from "@/services/user-service";
-import { getMenuBranches } from "@/lib/api-service";
+import { getMenuBranches } from "@/services/menu-service";
 import { showCreateSuccess, showError, showWarning } from "@/lib/sweetalert";
 import type { User, Branch } from "@/types/auth";
 import type { Role } from "@/types/role";
