+++ services/permission-service.ts
@@ -1,0 +1,13 @@
+// Types for API functions
+interface Permission {
+  id: number;
+  name: string;
+  description: string;
+}
+
+interface MenuPermission {
+  id: number;
+  name: string;
+  description: string;
+  menu?: string;
+}
