+++ services/role-service.ts
@@ -1,0 +1,63 @@
+import apiClient from "@/lib/api-client";
+import type { Role, CreateRoleData, UpdateRoleData } from "@/types/role";
+
+export interface RoleListResponse {
+  roles: Role[];
+  total: number;
+  page: number;
+  limit: number;
+  totalPages: number;
+}
+
+export interface GetRolesParams {
+  page?: number;
+  limit?: number;
+  search?: string;
+}
+
+// ดึงข้อมูลตำแหน่งทั้งหมด
+export async function getRoles(
+  params?: GetRolesParams
+): Promise<RoleListResponse> {
+  const queryParams = new URLSearchParams();
+
+  if (params?.page) queryParams.append("page", params.page.toString());
+  if (params?.limit) queryParams.append("limit", params.limit.toString());
+  if (params?.search) queryParams.append("search", params.search);
+
+  const url = `/api/v1/roles${
+    queryParams.toString() ? `?${queryParams.toString()}` : ""
+  }`;
+
+  const response = await apiClient.get<RoleListResponse>(url);
+  return response.data;
+}
+
+// ดึงข้อมูลตำแหน่งตาม ID
+export async function getRoleById(id: number): Promise<Role> {
+  const response = await apiClient.get<{ role: Role }>(`/api/v1/roles/${id}`);
+  return response.data.role;
+}
+
+// สร้างตำแหน่งใหม่
+export async function createRole(data: CreateRoleData): Promise<Role> {
+  const response = await apiClient.post<{ role: Role }>("/api/v1/roles", data);
+  return response.data.role;
+}
+
+// อัปเดตข้อมูลตำแหน่ง
+export async function updateRole(
+  id: number,
+  data: UpdateRoleData
+): Promise<Role> {
+  const response = await apiClient.put<{ role: Role }>(
+    `/api/v1/roles/${id}`,
+    data
+  );
+  return response.data.role;
+}
+
+// ลบตำแหน่ง
+export async function deleteRole(id: number): Promise<void> {
+  await apiClient.delete(`/api/v1/roles/${id}`);
+}
